service/requirements.txt
click==6.7
cryptography==0.6
Cython==0.25.2
Flask==0.12.2
funcsigs==1.0.2
gunicorn==19.7.1
h5py==2.7.0
itsdangerous==0.24
Jinja2==2.9.6
Keras==2.0.4
MarkupSafe==1.0
meld3==1.0.2
mock==2.0.0
numpy==1.13.0
olefile==0.44
pbr==3.0.1
Pillow==4.1.1
protobuf==3.3.0
PyYAML==3.12
scipy==0.19.0
six==1.10.0
supervisor==3.3.1
supervisor-stdout==0.1.1
tensorflow==1.1.0
Theano==0.9.0
virtualenv==15.1.0
Werkzeug==0.12.2
